% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tlp}
\alias{tlp}
\title{Tractatus Logico-Philosophicus}
\format{A data frame with 512 rows and 2 variables:
\describe{
  \item{id}{Wittgenstein's numbering as character strings}
  \item{text}{The respective proposition(s) as character strings}
}}
\source{
\url{https://archive.org/stream/tractatuslogicop05740gut/tloph10.txt}
}
\usage{
tlp
}
\description{
A dataset containing the full text of Ludwig Wittgenstein's Tractatus Logico-Philosophicus.
}
\keyword{datasets}

