% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tlp_intro}
\alias{tlp_intro}
\title{Tractatus Logico-Philosophicus (Introduction)}
\format{A character vector with length 39.}
\source{
\url{https://archive.org/stream/tractatuslogicop05740gut/tloph10.txt}
}
\usage{
tlp_intro
}
\description{
A dataset containing the introduction of Ludwig Wittgenstein's Tractatus Logico-Philosophicus.
}
\keyword{datasets}

